requirements.txt
anyio==4.3.0
blinker==1.7.0
docutils==0.20.1
feedgenerator==2.1.0
idna==3.6
Jinja2==3.1.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
pelican==4.9.1
pillow==10.2.0
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2024.1
rich==13.7.0
six==1.16.0
sniffio==1.3.0
Unidecode==1.3.8
watchfiles==0.21.0
